le data le phlowing

+++ src/generate/tileMapGen/tileMapGen.cpp
@@ -3,7 +3,7 @@
 #include <cstdint>
 #include <vector>
 
-std::vector<uint8_t> TileMapGen::GenerateNewTileMap(std::vector<uint8_t>& zIndex, TileSet& tileSet) {
-  std::vector<uint8_t> tileMap(zIndex.size(), 0); 
+std::vector<uint16_t> TileMapGen::GenerateNewTileMap(std::vector<uint8_t>& zIndex, TileSet& tileSet) {
+  std::vector<uint16_t> tileMap(zIndex.size(), 0); 
   return tileMap;
 }
