+++ src/generate/tileMapGen/tileMapGen.hpp
@@ -6,5 +6,5 @@
 #include <cstdint>
 
 namespace TileMapGen {
-  std::vector<uint8_t> GenerateNewTileMap(std::vector<uint8_t>& zIndex, TileSet& tileSet);
+  std::vector<uint16_t> GenerateNewTileMap(std::vector<uint8_t>& zIndex, TileSet& tileSet);
 }
